fix: validate per-node worker sizing at construction
